remove main.js duplicates, move it into src
